requirements.txt
pyspark==3.5.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
kafka-python==2.0.2
pyarrow==16.1.0  # Or any recent version >= 4.0.0